refactored db related logic in controllers: PaymentJsonController, PaymentMethodJsonController, RefundJsonController, ShippingFulfillmentJsonController - DEVE-45

+++ src/Controllers/RefundJsonController.php
@@ -3,12 +3,10 @@
 namespace Railroad\Ecommerce\Controllers;
 
 use Carbon\Carbon;
-use Doctrine\ORM\QueryBuilder;
 use Illuminate\Routing\Controller;
 use Railroad\Ecommerce\Entities\Order;
 use Railroad\Ecommerce\Entities\OrderItem;
 use Railroad\Ecommerce\Entities\OrderPayment;
-use Railroad\Ecommerce\Entities\Payment;
 use Railroad\Ecommerce\Entities\PaymentMethod;
 use Railroad\Ecommerce\Entities\Product;
 use Railroad\Ecommerce\Entities\Refund;
@@ -145,25 +143,7 @@ class RefundJsonController extends Controller
 
         $paymentId = $request->input('data.relationships.payment.data.id');
 
-        /**
-         * @var $qb QueryBuilder
-         */
-        $qb = $this->paymentRepository->createQueryBuilder('p');
-
-        $qb->select(['p', 'pm'])
-            ->join('p.paymentMethod', 'pm')
-            ->where(
-                $qb->expr()
-                    ->eq('p.id', ':id')
-            )
-            ->setParameter('id', $paymentId);
-
-        /**
-         * @var $payment Payment
-         */
-        $payment =
-            $qb->getQuery()
-                ->getOneOrNullResult();
+        $payment = $this->paymentRepository->getPaymentAndPaymentMethod($paymentId);
 
         /**
          * @var $paymentMethod PaymentMethod
@@ -212,26 +192,7 @@ class RefundJsonController extends Controller
 
         // cancel shipping fulfillment
 
-        /**
-         * @var $qb QueryBuilder
-         */
-        $qb = $this->orderPaymentRepository->createQueryBuilder('op');
-
-        $qb->select(['op', 'p'])
-            ->join('op.payment', 'p')
-            ->where(
-                $qb->expr()
-                    ->eq('op.payment', ':payment')
-            )
-            ->andWhere(
-                $qb->expr()
-                    ->isNull('p.deletedOn')
-            )
-            ->setParameter('payment', $payment);
-
-        $orderPayments =
-            $qb->getQuery()
-                ->getResult();
+        $orderPayments = $this->orderPaymentRepository->getByPayment($payment);
 
         $distinctOrders = [];
 
@@ -246,29 +207,8 @@ class RefundJsonController extends Controller
             $distinctOrders[$order->getId()] = $order;
         }
 
-        /**
-         * @var $qb QueryBuilder
-         */
-        $qb = $this->orderItemFulfillmentRepository->createQueryBuilder('oif');
-
-        $qb->where(
-                $qb->expr()
-                    ->in('oif.order', ':orders')
-            )
-            ->andWhere(
-                $qb->expr()
-                    ->eq('oif.status', ':status')
-            )
-            ->andWhere(
-                $qb->expr()
-                    ->isNull('oif.fulfilledOn')
-            )
-            ->setParameter('orders', array_values($distinctOrders))
-            ->setParameter('status', config('ecommerce.fulfillment_status_pending'));
-
-        $orderItemFulfillments =
-            $qb->getQuery()
-                ->getResult();
+        $orderItemFulfillments = $this->orderItemFulfillmentRepository
+                                        ->getByOrders(array_values($distinctOrders));
 
         foreach ($orderItemFulfillments as $orderItemFulfillment) {
             $this->entityManager->remove($orderItemFulfillment);
@@ -277,24 +217,18 @@ class RefundJsonController extends Controller
         // remove user products if full refund
         if ($refund->getPaymentAmount() == $refund->getRefundedAmount()) {
             if (count($orderPayments)) {
-                /**
-                 * @var $qb QueryBuilder
-                 */
-                $qb = $this->orderItemRepository->createQueryBuilder('oi');
 
-                $qb->select(['oi', 'p'])
-                    ->join('oi.product', 'p')
-                    ->where(
-                        $qb->expr()
-                            ->in('oi.order', ':orders')
-                    )
-                    ->setParameter('orders', array_values($distinctOrders));
-
-                $orderItems =
-                    $qb->getQuery()
-                        ->getResult();
+                $orderItems = $this->orderItemRepository->getByOrders(array_values($distinctOrders));
 
                 $distinctProducts = [];
+                $productsToUsersMap = [];
+                /*
+                // structure example
+                $productsToUsersMap = [
+                    product_id_1 => [user_id_1, user_id_2, user_id_3],
+                    product_id_2 => [user_id_1, user_id_4, user_id_5]
+                ];
+                */
 
                 /**
                  * @var $orderItem OrderItem
@@ -306,44 +240,32 @@ class RefundJsonController extends Controller
                     $product = $orderItem->getProduct();
 
                     $distinctProducts[$product->getId()] = $product;
+
+                    if ($orderItem->getOrder() && $orderItem->getOrder()->getUser() &&
+                        $orderItem->getOrder()->getUser()->getId()) {
+
+                        if (!isset($productsToUsersMap[$product->getId()])) {
+                            $productsToUsersMap[$product->getId()] = [];
+                        }
+
+                        $productsToUsersMap[$product->getId()][] = $orderItem->getOrder()->getUser()->getId();
+                    }
                 }
 
-                /**
-                 * @var $qb QueryBuilder
-                 */
-                $qb = $this->userProductRepository->createQueryBuilder('up');
-
-                $qb->where(
-                        $qb->expr()
-                            ->in('up.product', ':products')
-                    )
-                    ->setParameter('products', array_values($distinctProducts));
-
-                $userProducts =
-                    $qb->getQuery()
-                        ->getResult();
+                $userProducts = $this->userProductRepository->getByProducts(array_values($distinctProducts));
 
                 foreach ($userProducts as $userProduct) {
-                    $this->entityManager->remove($userProduct);
+                    $productId = $userProduct->getProduct()->getId();
+                    $userId = $userProduct->getUser()->getId();
+
+                    if (isset($productsToUsersMap[$productId]) && in_array($userId, $productsToUsersMap[$productId])) {
+                        $this->entityManager->remove($userProduct);
+                    }
                 }
             }
             else {
-                /**
-                 * @var $qb QueryBuilder
-                 */
-                $qb = $this->subscriptionPaymentRepository->createQueryBuilder('sp');
 
-                $qb->select(['sp', 's'])
-                    ->join('sp.subscription', 's')
-                    ->where(
-                        $qb->expr()
-                            ->eq('sp.payment', ':payment')
-                    )
-                    ->setParameter('payment', $payment);
-
-                $subscriptionPayments =
-                    $qb->getQuery()
-                        ->getResult();
+                $subscriptionPayments = $this->subscriptionPaymentRepository->getByPayment($payment);
 
                 /**
                  * @var $subscriptionPayment SubscriptionPayment
